Start on version 1.7.0

+++ src/main/common/maths.c
@@ -187,7 +187,7 @@ int scaleRange(int x, int srcMin, int srcMax, int destMin, int destMax) {
     return ((a / b) - (destMax - destMin)) + destMax;
 }
 
-int scaleRangef(float x, float srcMin, float srcMax, float destMin, float destMax) {
+float scaleRangef(float x, float srcMin, float srcMax, float destMin, float destMax) {
     float a = (destMax - destMin) * (x - srcMin);
     float b = srcMax - srcMin;
     return ((a / b) - (destMax - destMin)) + destMax;
